Rewrite pointer math namespace functions
- Add missing functions to the pointer math unit test
- Extend the alignment type from uint8_t to size_t

+++ include/nctl/FreeListAllocator.h
@@ -55,7 +55,7 @@ class DLL_PUBLIC FreeListAllocator : public IAllocator
 	struct Header
 	{
 		size_t size;
-		uint8_t adjustment;
+		size_t adjustment;
 	};
 
 	Block *freeBlocks_;
@@ -65,8 +65,8 @@ class DLL_PUBLIC FreeListAllocator : public IAllocator
 	FreeListAllocator(const FreeListAllocator &) = delete;
 	FreeListAllocator &operator=(const FreeListAllocator &) = delete;
 
-	static void *allocateImpl(IAllocator *allocator, size_t size, uint8_t alignment);
-	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, uint8_t alignment, size_t &oldSize);
+	static void *allocateImpl(IAllocator *allocator, size_t size, size_t alignment);
+	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, size_t alignment, size_t &oldSize);
 	static void deallocateImpl(IAllocator *allocator, void *ptr);
 };
