+++ include/nctl/ProxyAllocator.h
@@ -20,8 +20,8 @@ class DLL_PUBLIC ProxyAllocator : public IAllocator
 	ProxyAllocator(const ProxyAllocator &) = delete;
 	ProxyAllocator &operator=(const ProxyAllocator &) = delete;
 
-	static void *allocateImpl(IAllocator *allocator, size_t size, uint8_t alignment);
-	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, uint8_t alignment, size_t &oldSize);
+	static void *allocateImpl(IAllocator *allocator, size_t size, size_t alignment);
+	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, size_t alignment, size_t &oldSize);
 	static void deallocateImpl(IAllocator *allocator, void *ptr);
 };
