+++ include/nctl/StackAllocator.h
@@ -41,8 +41,8 @@ class DLL_PUBLIC StackAllocator : public IAllocator
 	StackAllocator(const StackAllocator &) = delete;
 	StackAllocator &operator=(const StackAllocator &) = delete;
 
-	static void *allocateImpl(IAllocator *allocator, size_t size, uint8_t alignment);
-	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, uint8_t alignment, size_t &oldSize);
+	static void *allocateImpl(IAllocator *allocator, size_t size, size_t alignment);
+	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, size_t alignment, size_t &oldSize);
 	static void deallocateImpl(IAllocator *allocator, void *ptr);
 };
